Crate stream_lines

Lined-oriented Rustlang Streams

Streams represent spools of results computed asyncronously. They are the async analog to Rustlang's Iterator type.

This crate represents a Stream transformer over Stream's of AsRef<[u8]> that result in Stream's of line-oriented values. Chunks of bytes are delimted by LF (\n) and optionally CRLF (\r\n) patterns.

You may find this crate useful if you are interfacing with a line-oriented protocol transmited over a network connection. Networks may split your protocol into uneven chunks of bytes. This crate buffers those chunks and yields a Stream that reconnects those lines.

Examples

The underlying abstraction is the Lines type which return's a stream over an arbitrarily converted type, but most applications will typically want to use the strings method which return's a Stream of Strings

extern crate futures;
extern crate stream_lines;
extern crate tokio;

use std::string::FromUtf8Error;

use futures::stream::iter_ok;
use futures::Stream;
use tokio::runtime::Runtime;

fn main() {
    let chunks = vec![
       "\nhello ",
        "world\n",
        "\n",
        "what a\nlovely",
        "\nday\n"
     ];
    let stream = iter_ok::<_, FromUtf8Error>(chunks);
    let print =
        stream_lines::strings(stream)
           .for_each(|line| Ok(println!("{}", line)));
    Runtime::new().expect("failed to initialize runtime").block_on(print).expect("failed to complete stream");
}

Structs

Lines
Converts a fused Stream of bytes into a line-oriented stream of a target type

Functions

strings
A lined oriented stream of Strings